chore: PR feedback
- expose getStorage to retrieve QueryStorage instead of createStorage
- use interfaces to describe the implementation being injected in
  plugins instead of actual implementation itself (Compass specific)
