+++ packages/compass-aggregations/src/modules/index.ts
@@ -31,7 +31,7 @@ import workspace from './workspace';
 import aggregationWorkspaceId from './aggregation-workspace-id';
 import type { ThunkAction, ThunkDispatch } from 'redux-thunk';
 import type { PipelineBuilder } from './pipeline-builder/pipeline-builder';
-import type { PipelineStorage } from '@mongodb-js/my-queries-storage';
+import type { PipelineStorage } from '@mongodb-js/my-queries-storage/provider';
 import focusMode from './focus-mode';
 import sidePanel from './side-panel';
 import collectionsFields from './collections-fields';
