+++ packages/compass-aggregations/src/modules/pipeline-builder/stage-editor.spec.ts
@@ -22,7 +22,7 @@ import {
 } from './stage-editor';
 import type { StageEditorState, StoreStage, Wizard } from './stage-editor';
 import reducer from '../';
-import { PipelineStorage } from '@mongodb-js/my-queries-storage';
+import { CompassPipelineStorage } from '@mongodb-js/my-queries-storage';
 import Sinon from 'sinon';
 import type Stage from './stage';
 import { mockDataService } from '../../../test/mocks/data-service';
@@ -117,7 +117,7 @@ function createStore({
         localAppRegistry: new AppRegistry(),
         atlasService: new AtlasService(),
         pipelineBuilder,
-        pipelineStorage: new PipelineStorage(),
+        pipelineStorage: new CompassPipelineStorage(),
         instance: {} as any,
         workspaces: {} as any,
         preferences,
