+++ packages/compass-aggregations/src/modules/pipeline-builder/text-editor-pipeline.spec.ts
@@ -11,7 +11,7 @@ import type { PipelineBuilderThunkDispatch } from '..';
 import reducer from '..';
 import Sinon from 'sinon';
 import { toggleAutoPreview } from '../auto-preview';
-import { PipelineStorage } from '@mongodb-js/my-queries-storage';
+import { CompassPipelineStorage } from '@mongodb-js/my-queries-storage';
 import { mockDataService } from '../../../test/mocks/data-service';
 
 function createStore(
@@ -47,7 +47,7 @@ function createStore(
       thunk.withExtraArgument({
         atlasService: new AtlasService(),
         pipelineBuilder,
-        pipelineStorage: new PipelineStorage(),
+        pipelineStorage: new CompassPipelineStorage(),
         instance: {} as any,
         workspaces: {} as any,
         preferences: {
