+++ packages/compass-aggregations/test/configure-store.ts
@@ -9,7 +9,7 @@ import type { DataService } from '../src/modules/data-service';
 import { ReadOnlyPreferenceAccess } from 'compass-preferences-model/provider';
 import { createNoopLoggerAndTelemetry } from '@mongodb-js/compass-logging/provider';
 
-import { PipelineStorage } from '@mongodb-js/my-queries-storage';
+import { CompassPipelineStorage } from '@mongodb-js/my-queries-storage';
 
 export default function configureStore(
   options: Partial<ConfigureStoreOptions> = {},
@@ -37,7 +37,7 @@ export default function configureStore(
       localAppRegistry: new AppRegistry(),
       workspaces: {} as any,
       logger: createNoopLoggerAndTelemetry(),
-      pipelineStorage: new PipelineStorage(),
+      pipelineStorage: new CompassPipelineStorage(),
       ...services,
     },
     createActivateHelpers()
