+++ packages/compass-schema/src/components/field.spec.tsx
@@ -13,8 +13,8 @@ import { configureActions } from '../actions';
 import Field, { shouldShowUnboundArrayInsight } from './field';
 import QueryBarPlugin from '@mongodb-js/compass-query-bar';
 import {
-  FavoriteQueryStorage,
-  RecentQueryStorage,
+  CompassFavoriteQueryStorage,
+  CompassRecentQueryStorage,
 } from '@mongodb-js/my-queries-storage';
 
 const MockQueryBarPlugin = QueryBarPlugin.withMockServices({
@@ -28,13 +28,13 @@ const MockQueryBarPlugin = QueryBarPlugin.withMockServices({
   },
   instance: { on() {}, removeListener() {} } as any,
   favoriteQueryStorageAccess: {
-    createStorage() {
-      return new FavoriteQueryStorage();
+    getStorage() {
+      return new CompassFavoriteQueryStorage();
     },
   },
   recentQueryStorageAccess: {
-    createStorage() {
-      return new RecentQueryStorage();
+    getStorage() {
+      return new CompassRecentQueryStorage();
     },
   },
 });
