+++ packages/my-queries-storage/src/index.ts
@@ -1,4 +1,7 @@
-export type { SavedPipeline } from './pipeline-storage';
-export { PipelineStorage } from './pipeline-storage';
-export { FavoriteQueryStorage, RecentQueryStorage } from './query-storage';
-export type { RecentQuery, FavoriteQuery } from './query-storage';
+export type { SavedPipeline } from './pipeline-storage-schema';
+export { CompassPipelineStorage } from './compass-pipeline-storage';
+export {
+  CompassFavoriteQueryStorage,
+  CompassRecentQueryStorage,
+} from './compass-query-storage';
+export type { RecentQuery, FavoriteQuery } from './query-storage-schema';
